requirements.txt
# CAPA AI Management System - Dependencies
# PT Bio Farma - Quality Assurance

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Agent & LLM
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.0.6
langchain-community>=0.0.20

# RAG & Embeddings
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-qdrant>=0.1.0
llama-index-vector-stores-postgres>=0.1.0
sentence-transformers>=2.2.0

# ML & XAI
xgboost>=2.0.0
shap>=0.44.0
mlflow>=2.10.0
scikit-learn>=1.3.0

# Vector Stores & Database
psycopg2-binary>=2.9.0
qdrant-client>=1.7.0

# Topic Modeling
umap-learn>=0.5.0
hdbscan>=0.8.33
bertopic>=0.16.0

# Validation & Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Security & Compliance
cryptography>=42.0.0

# HTTP & Async
httpx>=0.26.0

# LLM (optional - for OpenAI/Claude when not using vLLM)
litellm>=1.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
